Qualys VMDR unify (#33240)
* added fetch assets

* added test

* fixed cr comments

* unify

* removed the collector

* update redame, rn

* update rn. and added qualys-get-assets to yml

* fixed

* fixed mypy

* changed the limit

* update readme

* mypy

* added breaking Changes

* added since_datetime to test_module

* description

* fetch interval and dataset name
