--- Packs/qualys/Integrations/QualysEventCollector/command_examples
@@ -1,2 +1,0 @@
-!qualys-get-host-detections limit=2
-!qualys-get-activity-logs limit=2
