+++ Packs/qualys/Integrations/Qualysv2/Qualysv2.yml
@@ -1,4 +1,7 @@
 category: Vulnerability Management
+sectionOrder:
+- Connect
+- Collect
 commonfields:
   id: QualysV2
   version: -1
@@ -9,20 +12,50 @@ configuration:
   required: true
   type: 0
   additionalinfo: "When using asset-tag commands, the official documentation recommends that the SERVER URL parameter should be in the following format: `https://qualysapi.<tenant>.apps.qualys.com/<end-point>`. For more details see the integration documentation."
+  section: Connect
 - display: Username
   name: credentials
   required: true
   type: 9
+  section: Connect
 - display: Trust any certificate (not secure)
   name: insecure
   type: 8
   required: false
+  section: Connect
+  advanced: true
 - display: Use system proxy settings
   name: proxy
   type: 8
   required: false
-description: Qualys Vulnerability Management lets you create, run, fetch and manage reports, launch and manage vulnerability and compliance scans, and manage the host assets you want to scan for vulnerabilities and compliance.
-display: Qualys v2
+  section: Connect
+  advanced: true
+- defaultvalue: 3 days
+  section: Collect
+  display: Event first fetch time
+  name: first_fetch
+  additionalinfo: If "First event fetch time" is set for a long time ago, it may cause performance issues.
+  required: false
+  advanced: true
+  type: 0
+- defaultvalue: 10000
+  section: Collect
+  display: Event Fetch Limit
+  name: max_fetch_activity_logs
+  additionalinfo: Maximum number of events to fetch per fetch iteration.
+  advanced: true
+  required: false
+  type: 0
+- additionalinfo: The fetch interval. It is recommended to set it to 24 hours, and the minimum interval is 1 hour.
+  defaultvalue: 1440
+  display: Assets and Vulnerabilities Fetch Interval
+  name: assetsFetchInterval
+  type: 19
+  section: Collect
+  advanced: true
+  required: false
+description: Qualys Vulnerability Management lets you create, run, manage reports and to fetch Activity Logs, Assets and Vulnerabilities, launch and manage vulnerability and compliance scans, and manage the host assets you want to scan for vulnerabilities and compliance.
+display: Qualys VMDR
 name: QualysV2
 script:
   commands:
@@ -2760,7 +2793,38 @@ script:
     - name: csv_data
       description: The CSV data file containing the vCenter - ESXi mapping records that you want to purge.
       required: true
-  dockerimage: demisto/python3:3.10.13.87159
+  - arguments:
+    - auto: PREDEFINED
+      defaultValue: 'false'
+      description: If true, the command will create events, otherwise it will only display them.
+      name: should_push_events
+      predefined:
+      - 'true'
+      - 'false'
+      required: true
+    - description: Maximum number of results to return.
+      name: limit
+    - description: Date to return results from.
+      name: since_datetime
+    - description: Offset which events to return.
+      name: offset
+    description: Gets activity logs from Qualys.
+    name: qualys-get-events
+  - arguments:
+      - auto: PREDEFINED
+        defaultValue: 'false'
+        description: If true, the command will create assets, otherwise it will only display the amount of available assets.
+        name: should_push_assets
+        predefined:
+          - 'true'
+          - 'false'
+        required: true
+    description: Gets assets from Qualys.
+    name: qualys-get-assets
+  dockerimage: demisto/python3:3.10.14.90585
+  isfetchevents: true
+  isfetchassets: true
+  runonce: false
   script: ''
   subtype: python3
   type: python
