+++ Packs/qualys/ReleaseNotes/3_0_0.md
@@ -1,0 +1,11 @@
+
+#### Integrations
+##### Qualys VMDR
+- **Breaking Change**: The ***Qualys Event Collector*** was removed from this pack. To fetch events and assets from Qualys, use the ***Qualys VMDR*** integration instead.
+- Updated the Docker image to: *demisto/python3:3.10.14.90585*.
+- Renamed the integration from ***Qualys v2*** to ***Qualys VMDR***.
+- Added the  following commands:
+  - ***fetch-events***
+  - ***fetch-assets*** 
+  - ***qualys-get-events***
+  - ***qualys-get-assets***
